new wallpapers

+++ config/waybar/scripts/launch.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-killall -9 waybar
-
-waybar &
+if pgrep -x waybar > /dev/null; then
+    killall -9 waybar
+else
+    waybar &
+fi
